Add teachers timetable, Fix message overflow on a long weeks, Add settings menu, Add premium prototype.

+++ sql/callback_messages.sql
@@ -1,7 +1,7 @@
 create table callback_messages
 (
-    message_id int null,
-    chat_id    int null,
-    day        int null,
-    month      int null
+    message_id int    null,
+    chat_id    bigint null,
+    day        int    null,
+    month      int    null
 );
